LayoutTests/imported/w3c/web-platform-tests/editing/plaintext-only/paste.https_white-space=pre-wrap-expected.txt

abc
def
AB

PASS Pasting text in <b>
PASS Pasting text in <span>
PASS Pasting text into editable <b>
PASS Pasting text in <i> into editable <b>
PASS Pasting 2 paragraphs
PASS Pasting 2 paragraphs into <b>
PASS Pasting 2 paragraphs whose text is bold
FAIL Pasting text in <b>: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting text in <b>: pasted result assert_equals: <b> should not be pasted expected "AabcB" but got "AB"
FAIL Pasting text in <span>: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting text in <span>: pasted result assert_equals: <span> should not be pasted expected "AabcB" but got "AB"
FAIL Pasting text into editable <b>: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting text into editable <b>: pasted result assert_equals: text should be inserted into the editable <b> expected "<b>AabcB</b>" but got "<b>AB</b>"
FAIL Pasting text in <i> into editable <b>: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting text in <i> into editable <b>: pasted result assert_equals: text should be inserted into the editable <b> without copied <i> expected "<b>AabcB</b>" but got "<b>AB</b>"
FAIL Pasting 2 paragraphs: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting 2 paragraphs: pasted result assert_in_array: Each paragraph should be pasted as a line value "AB" not in array ["Aabc\ndefB", "A\nabc\ndef\nB"]
FAIL Pasting 2 paragraphs into <b>: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting 2 paragraphs into <b>: pasted result assert_in_array: Each paragraph should be pasted as a line value "<b>AB</b>" not in array ["<b>Aabc\ndefB</b>", "<b>A\nabc\ndef\nB</b>"]
FAIL Pasting 2 paragraphs whose text is bold: beforeinput assert_equals: inputType should be "insertFromPaste" expected (string) "insertFromPaste" but got (undefined) undefined
FAIL Pasting 2 paragraphs whose text is bold: pasted result assert_in_array: Each paragraph should be pasted as a line value "AB" not in array ["Aabc\ndefB", "A\nabc\ndef\nB"]
